+++ karma.conf.js
@@ -1,3 +1,11 @@
+const {
+  BROWSER,
+  VIRTUALBOX_EDGE_UUID,
+  VIRTUALBOX_IE11_UUID,
+  ANDROID_HOME,
+  ANDROID_AVD_NAME
+} = process.env;
+
 const staticConfig = {
   basePath: '',
   frameworks: ['mocha', 'sinon', 'source-map-support'],
@@ -8,11 +16,12 @@ const staticConfig = {
   proxies: {
     '/api': 'http://localhost:9877'
   },
+  hostname: (BROWSER.includes('Android') ? '10.0.2.2' : 'localhost'),
   reporters: ['mocha', 'coverage'],
   port: 9876,
   colors: true,
   autoWatch: true,
-  browsers: [process.env.BROWSER || 'ChromeHeadless'],
+  browsers: [BROWSER || 'ChromeHeadless'],
   singleRun: true,
   concurrency: Infinity,
   browserDisconnectTimeout: 800000,
@@ -31,12 +40,17 @@ const staticConfig = {
     VirtualBoxEdgeWin10: {
       base: 'VirtualBoxEdge',
       keepAlive: true,
-      uuid: process.env.VIRTUALBOX_EDGE_UUID
+      uuid: VIRTUALBOX_EDGE_UUID
     },
     VirtualBoxIE11Win7: {
       base: 'VirtualBoxIE11',
       keepAlive: true,
-      uuid: process.env.VIRTUALBOX_IE11_UUID
+      uuid: VIRTUALBOX_IE11_UUID
+    },
+    Android10: {
+      base: 'AndroidEmulator',
+      avdName: ANDROID_AVD_NAME || 'Pixel_3_API_29',
+      sdkHome: ANDROID_HOME
     }
   },
   client: {